# docs/swagger.yaml
openapi: 3.0.3
info:
  title: TeamSupportPro API
  description: REST API for TeamSupportPro — a team-based IT support ticket management system.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: team_support_session
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [Admin, Staff] }
        teamId: { type: string }
        teamName: { type: string }
        teamCode: { type: string }
        teamAccent: { type: string }
        picture: { type: string }
    Error:
      type: object
      properties:
        error: { type: string }
...
security:
  - cookieAuth: []
  - apiKeyHeader: []
  - bearerAuth: []
paths:
...
  /auth/oidc/callback:
    get:
      summary: OIDC callback
      description: Handles the OIDC provider callback, exchanges code for tokens, and sets session cookie.
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        "302":
          description: Redirect to client URL with session cookie set
  /api/auth/me:
    get:
      summary: Get current authenticated user
      responses:
        "200":
          description: Authenticated user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated: { type: boolean }
                  user: { $ref: '#/components/schemas/AuthUser' }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/register:
    post:
      summary: Register a local account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, minLength: 2 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Account created and session started
        "400":
          description: Validation error
        "409":
          description: Email already exists
  /api/auth/local/login:
    post:
      summary: Login with local account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
        "400":
          description: Invalid email format
        "401":
          description: Invalid credentials
...